+++ packages/logger/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @appbuildersph/logger
 
+## 0.0.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @appbuildersph/shared@0.1.0
+
 ## 0.0.3
 
 ### Patch Changes